datalake_cdk/pyproject.toml
[tool.poetry]
name = "datalake-cdk"
version = "0.1.0"
description = ""
authors = ["Martijn Sturm <martijn.sturm@cloudnation.nl>"]
readme = "README.md"
packages = [{include = "datalake_cdk"}]

[tool.poetry.dependencies]
python = "^3.10"
datalake-lib = {path = "../datalake_lib", develop = true}
aws-cdk-lib = "^2.73.0"
constructs = "^10.1.306"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
